+++ mmms-backend/src/main/java/com/sep/mmms_backend/controller/MeetingController.java
@@ -8,11 +8,9 @@ import com.sep.mmms_backend.response.ResponseMessages;
 import com.sep.mmms_backend.service.MeetingService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -26,43 +24,39 @@ public class MeetingController {
         this.meetingService = meetingService;
    }
 
-    @PostMapping("/createMeeting")
-    public ResponseEntity<Response> createMeeting(@RequestBody @Valid Meeting meeting, Errors errors) {
 
-       //if meetingId is populated in the request body, then the operation should be 'save' not 'update'
-       if(meeting.getMeetingId() != 0) {
-           return ResponseEntity.badRequest().body(new Response(ResponseMessages.ROUTE_CREATE_MEETING_MISUSED));
-       }
-
-        //Validation failed:
-        if(errors.hasErrors()) {
-            HashMap<String, ArrayList<String>> errorMessages = new HashMap<>();
-
-            errors.getFieldErrors().forEach(
-                    config-> {
-                        if(errorMessages.containsKey(config.getField())) {
-                              errorMessages.get(config.getField()).add(config.getDefaultMessage());
-                        } else {
-                            ArrayList<String> list = new ArrayList<>();
-                            list.add(config.getDefaultMessage());
-                            errorMessages.put(config.getField(), list);
-                        }
-                    }
-            );
-
-            return ResponseEntity.badRequest().body(new Response(ResponseMessages.MEETING_CREATION_FAILED, errorMessages));
+    /*
+    {
+      "meetingHeldDate": "2025-07-09",
+      "meetingHeldPlace": "Kathmandu, Nepal",
+      "meetingHeldTime": "14:30:00",
+      "attendees": [
+        {
+          "memberId": 101
+        },
+        {
+          "memberId": 102
+        },
+        {
+          "memberId": 103
         }
-
-        try {
-           Meeting savedMeeting = this.meetingService.saveNewMeeting(meeting);
-        } catch(MeetingAlreadyExistsException e) {
-            return ResponseEntity.badRequest().body(new Response(ResponseMessages.ROUTE_CREATE_MEETING_MISUSED));
+      ],
+      "decisions": [
+        {
+          "decision": "Approved annual budget."
+        },
+        {
+          "decision": "Scheduled next meeting for August."
         }
-
-        return ResponseEntity.ok().body(new Response(ResponseMessages.MEETING_CREATION_SUCCESSFUL));
+      ]
     }
+    */
 
-
+    @PostMapping("committee/createMeeting")
+    public ResponseEntity<Response> createMeeting(@RequestBody(required = true) Meeting meeting, @RequestParam(required=true) int committeeId, Authentication authentication) {
+       Meeting savedMeeting =  meetingService.saveNewMeeting(meeting, committeeId, authentication.getName());
+       return ResponseEntity.ok(new Response(savedMeeting));
+    }
 
 
     @PostMapping("/updateMeeting")