+++ mmms-backend/src/main/java/com/sep/mmms_backend/exceptions/ExceptionMessages.java
@@ -14,7 +14,8 @@ public enum ExceptionMessages {
     VALIDATION_FAILED("Validation Failed"),
     COMMITTEE_DOES_NOT_EXIST("The specified committee does not exist"),
     NO_VALID_MEMBERSHIP("No valid membership"),
-    COMMITTEE_NOT_ACCESSIBLE("Specified committee is not accessible")
+    COMMITTEE_NOT_ACCESSIBLE("Specified committee is not accessible"),
+    MEMBER_NOT_IN_COMMITTEE("The specified member is not associated with the committee"),
     ;