+++ mmms-backend/src/main/java/com/sep/mmms_backend/service/AppUserService.java
@@ -1,7 +1,6 @@
 package com.sep.mmms_backend.service;
 
 import com.sep.mmms_backend.entity.AppUser;
-import com.sep.mmms_backend.entity.Committee;
 import com.sep.mmms_backend.exceptions.*;
 import com.sep.mmms_backend.repository.AppUserRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +10,6 @@ import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.Validator;
 
-import java.util.List;
 import java.util.Optional;
 
 @Slf4j